+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+* Add CI specific git info class
+
 ## 0.0.2
 * Only match exact git tags